% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censorandtrim.R
\name{censor.Surv}
\alias{censor.Surv}
\title{Censor and Trim}
\usage{
censor.Surv(x, censor.times)
}
\arguments{
\item{x}{Surv object}

\item{censor.times}{time of  (right) censoring, or vector of times of censoring}
}
\value{
Censored / trimmed Surv object
}
\description{
Internal functions for right-censoring and left-trimming survival data.
}
