% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cycloSurv.R
\name{create_cycloSurv}
\alias{create_cycloSurv}
\title{Create an object that can be fitted with a given period from survival (time of death or censoring) data}
\usage{
create_cycloSurv(start, end, event, t0 = NULL, period,
  timeunits = "days")
}
\arguments{
\item{start}{a vector measuring time of birth (as a POSIXct, numeric, or Date)}

\item{end}{a vector measuring time of death or censoring (as a POSIXct, numeric, or Date)}

\item{event}{a vector of booleans (0 = alive, 1 = dead) detailing the status of the observation}

\item{t0}{reference time for event times.  By default, \code{t0} is set to January 1 of the first year of observations (if times are POSIX).}

\item{period}{length of one period in the input data}

\item{timeunits}{units that dates are inputted in if dates are being used}

\item{data}{data.frame object containing start, end and event data}
}
\value{
a Surv object with an attribute "period" that reads the period.
}
\description{
Create an object that can be fitted with a given period from survival (time of death or censoring) data
}
\examples{
startTimes = as.Date(origin = "2010-01-01", 
                    c(0, 0, 0, 50, 0, 50, 100, 150, 0, 100)) #in days
endTimes = as.Date(origin = "2010-01-01", 
                  c(50, 50, 100, 150, 150, 200, 200, 250, 350, 500)) #in days
censored = c(1, 1, 0, 1, 1, 0, 1, 0, 0, 0)
period = 365
morts = create_cycloSurv(start = startTimes, end = endTimes, 
                       event = censored, period = period)
}
