% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCycloSurv.R
\name{createCycloSurv}
\alias{createCycloSurv}
\title{Create an object that can be fitted with a given period from survival (time of death or censoring) data}
\usage{
createCycloSurv(data, censoring, period)
}
\arguments{
\item{censoring}{a vector of booleans (0 = alive, 1 = dead) detailing}

\item{period}{length of one period in "data"}

\item{times}{a vector of Surv objects measuring time of death or censoring}
}
\value{
a Surv object with an attribute "period" that reads the period.
}
\description{
Create an object that can be fitted with a given period from survival (time of death or censoring) data
}
