\name{cyclomort-package}
\alias{cyclomort-package}
\alias{cyclomort}
\docType{package}
\title{
\packageTitle{cyclomort}
}
\description{
\packageDescription{cyclomort}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cyclomort}
\packageIndices{cyclomort}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{cyclomort}

Maintainer: \packageMaintainer{cyclomort}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
