% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwrappedCauchy.R
\name{DwrappedCauchy}
\alias{DwrappedCauchy}
\title{Wrapped Cauchy distribution}
\usage{
DwrappedCauchy(theta, mu, rho)
}
\arguments{
\item{theta}{angle}

\item{mu}{mean parameter}

\item{rho}{concentration parameter (between -1 and 1), equivalent to the E(cos(theta))}
}
\description{
Very basic Wrapped Cauchy distribution function
}
\examples{
plot(DwrappedCauchy(seq(-pi, pi, .01), 0, 0.7), type = "l")
}
