% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwrappedMultiCauchy.R
\name{DwrappedMultiCauchy}
\alias{DwrappedMultiCauchy}
\title{Mixed Cauchy distribution}
\usage{
DwrappedMultiCauchy(theta, mus, rhos, weights, A = 1)
}
\arguments{
\item{theta}{angle}

\item{mus}{k-vector of mean parameters}

\item{rhos}{k-vector of concentration parameters}

\item{weights}{(k-1)-vector of weights - the sum must be less than 1}

\item{A}{overall magnitude}
}
\value{
the value of the p.d.f. of the mixed distribution
}
\description{
Very basic Wrapped Cauchy shaped function. Incorporates any number of mixed
distributions in order to simulate a potential function with a variety of differently
shaped peaks over a finite period.
}
\examples{
plot(seq(-pi, pi, .01), 
DwrappedMultiCauchy(seq(-pi, pi, .01), 
A = 1, 
mus = c(-1,1), 
rhos = c(0.7,0.5), 
weights = 0.7), type = "l")
}
\seealso{
DwrappedCauchy
}
