% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cyclomort.R
\name{fit_cyclomort}
\alias{fit_cyclomort}
\title{Get maximum likelihood estimates for periodic mortality data. Uses the optim
function along with a named vector of initial guesses and returns the parameter
estimates.}
\usage{
fit_cyclomort(T, p0, dt)
}
\arguments{
\item{T}{set of Surv objects representing time of death or censorship}

\item{p0}{set of initial guesses; a named vector with "A", "peak", "rho" and "weight" names}

\item{dt}{interval for plots as well as precision of random samples}

\item{period}{period of one mortality cycle}
}
\value{
parameter estimates for weights, rhos, peaks and A
}
\description{
Get maximum likelihood estimates for periodic mortality data. Uses the optim
function along with a named vector of initial guesses and returns the parameter
estimates.
}
\examples{
T.morts1 <- simPeriodicMorts(300, period = 1, 
                             peaks = c(0.25, 0.75), 
                             rhos = c(0.8, 0.5), 
                             weights = 0.3, dt = .01, 
                             A = .02, plotme = FALSE)

p0 = c(A = 0.05, peak1 = 0.25, peak2 = 0.75, rho1 = 0.5, rho2 = 0.5, weight1 = 0.5)
##MLE for parameters based on simulated data
fits = fit_cyclomort(T.morts1, p0, dt = .01, period = 1)
getCIs(fits)
##Actual parameter values from simulated data
attributes(T.morts1)
}
