% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cyclomort.R
\name{generateInitialParameterEstimate}
\alias{generateInitialParameterEstimate}
\title{Produce initial parameter estimates based on mortality data}
\usage{
generateInitialParameterEstimate(T, n, null_fits)
}
\arguments{
\item{T}{set of Surv objects representing time of death or censorship}

\item{n}{expected number of seasons within a period}

\item{null_fits}{original estimate for mortality rate assuming constant hazard function}
}
\value{
a named vector listing intial guesses for parameter values, to be used in the fitting process
}
\description{
Uses a basic flexsurvreg exponential mortality model to find the average hazard value, and fits a mixed normal distribution model to estimate the peaks, season durations, and weight distributions for the model. These estimates are not meant to be fully accurate but instead are meant to be good initial guesses for the fit_cyclomort function.
}
