% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cyclomort.R
\name{getCIs}
\alias{getCIs}
\title{Get 95% confidence intervals for each parameter that is estimated by the MLE}
\usage{
getCIs(fit)
}
\arguments{
\item{fit}{returned value from call to "optim" containing parameter estimates and Hessian matrix}
}
\value{
confidence intervals for each parameter in the estimate
}
\description{
Get 95% confidence intervals for each parameter that is estimated by the MLE
}
