% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cyclomort.R
\name{getSeasonLength}
\alias{getSeasonLength}
\title{Gets the season length in periods based on the concentration parameter rho}
\usage{
getSeasonLength(rho)
}
\arguments{
\item{rho}{concentration parameter for wrapped Cauchy distribution between 0 and 1}
}
\value{
length of season assuming period length of 1
}
\description{
Gets the season length in periods based on the concentration parameter rho
}
\examples{
plot(Vectorize(getSeasonLength)(seq(0.05, 0.95, 0.05)), type = "l")
}
