% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_cmfit.R
\name{logLik.cmfit}
\alias{logLik.cmfit}
\title{Determine log-likelihood for a cmfit object}
\usage{
\method{logLik}{cmfit}(x)
}
\arguments{
\item{x}{a cmfit object}
}
\value{
log-likelihod for the cmfit object to be used for comparison with other fits
}
\description{
Determine log-likelihood for a cmfit object
}
\examples{
T.morts1 <- simPeriodicMorts(300, period = 1, 
                             peaks = c(0.25, 0.75), 
                             rhos = c(0.8, 0.5), 
                             weights = 0.3, dt = .01, 
                             A = .02, plotme = FALSE)

p0 = c(A = 0.05, peak1 = 0.25, peak2 = 0.75, rho1 = 0.5, rho2 = 0.5, weight1 = 0.5)
##MLE for parameters based on simulated data
fits = fit_cyclomort(T.morts1, p0, dt = .01)
print(fits)
##Actual parameter values from simulated data
attributes(T.morts1)
##Measures of model accuracy
AIC(fits)
logLik(fits)
}
