% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{loglike}
\alias{loglike}
\title{Obtain log-likelihood value from a data set given a set of parameter values}
\usage{
loglike(T, A, p, r, w, dt, period)
}
\arguments{
\item{T}{times of death or censoring as Surv objects}

\item{A}{average annual hazard}

\item{p}{k-vector of peaks}

\item{r}{k-vector of rhos (concentration parameters)}

\item{w}{(k-1)-vector of weights}

\item{dt}{interval for plots as well as precision of random samples}

\item{period}{period of one mortality cycle}
}
\value{
the maximum likelihood value for this set of data
}
\description{
Obtain log-likelihood value from a data set given a set of parameter values
}
\examples{
require(fields);

T.morts1 <- simPeriodicMorts(300, period = 1, 
                             peaks = c(0.25, 0.75), 
                             rhos = c(0.8, 0.5), 
                             weights = 0.3, dt = .01, 
                             A = .02, plotme = FALSE);

peaks = seq(0, 1, length = 30);
##Is there a way to vectorize this function more easily than what I do here?
ll.matrix = matrix(nrow = 30, ncol = 30, 0);
for (i in 1:30) {
  for (j in 1:30) {
    ll.matrix[i, j] = loglike(T.morts1, A = 0.02, p = c(peaks[i], peaks[j]), r = c(0.8, 0.5), w = c(0.3), 0.01, period = 1);
  }
}

image.plot(peaks, peaks, ll.matrix)
}
