% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{loglike_optim}
\alias{loglike_optim}
\title{Log-likelihood function that is useable for the optim command}
\usage{
loglike_optim(pars, T)
}
\arguments{
\item{pars}{named vector or list including "weight", "peak", and "duration" parameters for the appropriate number of seasons}

\item{T}{times of death or censoring as Surv objects}
}
\value{
likelihood value given named vector of parameters as well as set of observations
}
\description{
Log-likelihood function that is useable for the optim command
}
