% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cyclomort.R
\name{model_cyclomort}
\alias{model_cyclomort}
\title{Compare model fits based on a factor - infer the effect of a categorical variable on 
the parameter estimates produced by fit_cyclomort.}
\usage{
model_cyclomort(formula, data, n.seasons = 2)
}
\arguments{
\item{formula}{formula object used for identifying different classes}

\item{data}{a data frame containing a cycloSurv object detailing mortalities for a set of observations and a factor identifying the value of a categorical variable for each observation}

\item{n.seasons}{number of seasons to fit model to}
}
\value{
table comparing outputs from null (factor has no effect on mortality and they are all
in the same group) model to multi-factor model using AIC, log-likelihood and likelihood ratio test
}
\description{
Compare model fits based on a factor - infer the effect of a categorical variable on 
the parameter estimates produced by fit_cyclomort.
}
