% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.r
\docType{data}
\name{nwt_morts}
\alias{nwt_morts}
\title{Mortality data for Northwest territory boreal woodland caribou}
\format{Data frame containing movements of roe deer with the following columns:
\describe{
  \item{id}{ID of animal}
  \item{region}{Subpopulation marker - for comparisons}
  \item{start}{Date of beginning of collaring}
  \item{end}{Date of death or censoring}
  \item{Fate}{One of "Accident", "Harvest", "Non-predation", "Pradation", "Unknown",  only for mortality events.}
  \item{Status}{"Mort" or "Cens" (right-censoring of data)}
  \item{interval}{days}}}
\usage{
data(nwt_morts)
}
\description{
Describing the data.
}
\examples{
data(nwt_morts)
require(ggplot2)
ggplot(nwt_morts \%>\% arrange(start) \%>\% mutate(id = factor(id, levels = id)),
aes(x = start, y = id, col = status)) + 
  geom_errorbarh(aes(xmin = start, xmax = end))

}
\keyword{data}
