% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmfactorfit.R
\name{plot.cmfactorfit}
\alias{plot.cmfactorfit}
\title{Plot cmfactorfit objects}
\usage{
\method{plot}{cmfactorfit}(x, fit = "both", colors = NULL,
  legend = TRUE, ...)
}
\arguments{
\item{x}{a cmfactorfit object}

\item{fit}{a character (either "null", "alt", or "both") that dictates what fits will be plotted}

\item{colors}{vector of colors}

\item{legend}{whether to automatically make a legend}

\item{...}{additional parameters to pass to the \code{\link{plot.cmfit}} function. Perhaps most usefully: lowering the default nreps (e.g. to 1000) makes plotting much faster.}
}
\value{
a plot comparing the estimates from the null model with the individual estimates from each factor level
}
\description{
Plot cmfactorfit objects
}
\examples{
T <- simulate_cycloMort(300, period = 1, peaks = c(0.3, 0.8), 
                      durations = c(0.15, 0.20), weights = c(3, 2)/5, 
                      meanhazard = 1, plotme = FALSE, max.periods = 6)
T.fit <- fit_cyclomort(T, n.seasons = 2)
predictions = predict(T.fit, t = 0.5, CI = TRUE, type = "hazard")

plot(T.fit)
}
