% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmfit.R
\name{plot.cmfit}
\alias{plot.cmfit}
\title{Plot cmfit objects}
\usage{
\method{plot}{cmfit}(x, CI = TRUE, histogram = TRUE, add = FALSE,
  nreps = 10000, hazcolor = "black", alpha = 0.5, ...)
}
\arguments{
\item{x}{a cmfit object}

\item{CI}{boolean variable determining whether confidence intervals are included in the hazard function estimate}

\item{histogram}{boolean variable determining whether a histogram will be included with the hazard}

\item{add}{boolean variable determining whether the plot will be added to an existing plot}

\item{nreps}{number of repetitions of random parameter sampling; used to develop confidence intervals}

\item{hazcolor}{color of lines for hazard function and confidence intervals}
}
\value{
a plot comparing the estimated mortality curve (based on parameter estimates)
and the actual results (as a histogram).
}
\description{
Plot cmfit objects
}
\examples{
T.morts1 <- simPeriodicMorts(1000, period = 365, 
                             meanhazard = 0.3 / 365, 
                             peaks = c(0.25 * 365, 0.75 * 365), 
                             durations = c(0.3 * 365, 0.1 * 365), 
                             weights = 0.7, 
                             plotme = FALSE)

##MLE for parameters based on simulated data
fits = fit_cyclomort(T.morts1, n.seasons = 2)
print(fits)
plot(fits)
##Actual parameter values from simulated data
attributes(T.morts1)
##Measures of model accuracy
AIC(fits)
logLik(fits)
}
