% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmfit.R
\name{plotHazard}
\alias{plotHazard}
\title{Plots the hazard curve for one period given either a cmfit object or a set of parameters}
\usage{
plotHazard(cmfit = NULL, pars, period, ...)
}
\description{
to do!!
}
