% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cmfit.R
\name{predict.cmfit}
\alias{predict.cmfit}
\title{Prediction method for cyclomort fits}
\usage{
\method{predict}{cmfit}(x, t = seq(0, x$period, length = 100),
  CI = FALSE, nreps = 10000)
}
\arguments{
\item{x}{a cmfit object}

\item{t}{times for prediction}

\item{CI}{whether or not to compute 95% confidence intervals}

\item{nreps}{number of samples drawn to generate confidence intervals.  The default 10^4 is more than enough, but is a little sluggish. 10^3 gives slightly rougher intervals, but is very fast}
}
\description{
Obtain predictions and - optionally - confidence intervals for the hazard function from a fitted periodic hazard functon.
}
