% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cmfit.R
\name{print.cmfit}
\alias{print.cmfit}
\title{Print a short summary of cmfit (parameter estimates for periodic mortality curves) objects}
\usage{
\method{print}{cmfit}(x, maxDigits = 4)
}
\arguments{
\item{x}{a cmfit object}

\item{maxDigits}{number of digits to round to}
}
\value{
a short summary of the estimates for each parameter along with confidence intervals and AIC
}
\description{
Print a short summary of cmfit (parameter estimates for periodic mortality curves) objects
}
\examples{
T.morts1 <- simPeriodicMorts(300, period = 1, 
                             peaks = c(0.25, 0.75), 
                             rhos = c(0.8, 0.5), 
                             weights = 0.3, dt = .01, 
                             A = .02, plotme = FALSE)

p0 = c(A = 0.05, peak1 = 0.25, peak2 = 0.75, rho1 = 0.5, rho2 = 0.5, weight1 = 0.5)
##MLE for parameters based on simulated data
fits = fit_cyclomort(T.morts1, p0, dt = .01)
print(fits)
##Actual parameter values from simulated data
attributes(T.morts1)
##Measures of model accuracy
AIC(fits)
logLik(fits)
}
