% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cmfitlist.R
\name{print.cmfitlist}
\alias{print.cmfitlist}
\title{Display cmfitlist objects as text}
\usage{
\method{print}{cmfitlist}(x)
}
\arguments{
\item{x}{a cmfitlist object}
}
\value{
a table comparing the different periodic mortality models, comparing them using AIC as well as the parameter estimates
}
\description{
Display cmfitlist objects as text
}
\examples{
T.morts1 <- simPeriodicMorts(300, period = 1, 
                             peaks = c(0.25, 0.75), 
                             rhos = c(0.7, 0.7), 
                             weights = 0.2, dt = .01, 
                             A = .02, plotme = FALSE)

fits = getIdealPeakFit(T = T.morts1, dt = 0.01)
print(fits)
}
