% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwrappedCauchy.R
\name{pwrpCauchy}
\alias{pwrpCauchy}
\title{CDF values for wrapped Cauchy distribution}
\usage{
pwrpCauchy(q, mu, rho)
}
\arguments{
\item{q}{desired value of the point, ranging from -pi to pi}

\item{mu}{mean or peak parameter}

\item{rho}{concentration parameter (between -1 and 1), equivalent to the E(cos(theta))}
}
\value{
CDF value for the given point, ranging from 0 to 1
}
\description{
Derive an integral to obtain the CDF value of wrapped Cauchy distribution at a point
}
\examples{
plot(Vectorize(pwrpCauchy, vectorization.args = "p")(seq(-pi, pi, 0.01), 0, 0.7), type = "l")
}
