% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwrappedCauchy.R
\name{qwrpCauchy}
\alias{qwrpCauchy}
\title{CDF quantiles for wrapped Cauchy distribution}
\usage{
qwrpCauchy(p, mu, rho, tol = 1e-05)
}
\arguments{
\item{p}{desired quantile (ranging from 0 to 1)}

\item{mu}{mean or peak parameter}

\item{rho}{concentration parameter (between -1 and 1), equivalent to the E(cos(theta))}
}
\value{
quantile value for the distribution on the scale of -pi to pi
}
\description{
Obtains quantile values for the wrapped Cauchy CDF
}
\examples{
plot(Vectorize(qwrpCauchy, vectorization.args = "p")(seq(0, 1, 0.01), 0, 0.7), type = "l")
}
