% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.r
\docType{data}
\name{seasonalsex}
\alias{seasonalsex}
\title{#' Mortality data for Northwest territory boreal woodland caribou.
#'
#' 
#' @usage 
#' data(nwt_morts)
#' 
#' @format Data frame with 370 rows and the following columns:
#' \describe{
#'   \item{id}{ID of animal}
#'   \item{region}{Subpopulation marker - for comparisons}
#'   \item{start}{Date of beginning of collaring}
#'   \item{end}{Date of death or censoring}
#'   \item{status}{"Mort" or "Cens" (dead or censored)}
#'   }
#' 
#' @examples 
#' data(nwt_morts)
#' require(ggplot2)
#' ggplot(nwt_morts %>% arrange(start) %>% mutate(id = factor(id, levels = id)),
#' aes(x = start, y = id, col = status)) + 
#'   geom_errorbarh(aes(xmin = start, xmax = end))
#' 
#' @source Government of Northwest Territories, Canada
#' @keywords data
"nwt_morts"
Simulated data of seasonal mortality data for two sex groups}
\format{Simulated data of single-season mortalities for two sex groups:
\describe{
  \item{sex}{female (F) or male (M)}
  \item{T}{\code{\link{cycloSurv}} object of (censored) survival data}
}}
\usage{
data(seasonalsex)
}
\description{
See examples below for the process of simulating and visualizing these data using \code{\link{simPeriodicMorts}}, and an example of analyzing these data with \code{\link{factorfit_cyclomort}}.
}
\examples{
# Example of simulating multi-factor data

eval <- FALSE; if(eval){
  n <- 100
  T.male = simPeriodicMorts(n, period = 1, meanhazard = 0.3, peaks = .25, durations = .3)
  T.female = simPeriodicMorts(n, period = 1, meanhazard = 0.3, peaks = .75, durations = .3)
  T <- with(rbind(T.male, T.female) \%>\% data.frame,
            createCycloSurv(start = start, end = stop, 
                            event = status, period = 1))
  seasonalsex <- data.frame( sex = rep(c("M","F"), each = n), T = T)
}


# load and visualize simulated sex-specific survival data

data("seasonalsex")

seasonsex.df <- cbind(seasonalsex, as.matrix(seasonalsex$T) \%>\% as.data.frame) \%>\%
  arrange(sex,stop) \%>\% mutate(id = 1:length(start) \%>\% factor, 
                               status = c("Dead", "Censored")[2-status])
ggplot(seasonsex.df, aes(x = start, y = id, col = sex)) + 
  geom_errorbarh(aes(xmin = start, xmax = stop, lty = status)) + 
    ggtitle("Simualted sex-specific mortality data")

ss.dead <- seasonsex.df \%>\% subset(status == "Dead") \%>\% mutate(time = (stop - floor(stop))) 
ggplot(ss.dead, aes(time, y=..density.., fill = sex)) + geom_density(alpha = 0.1) + 
  geom_histogram(alpha=0.5, position="identity")  + 
    ggtitle("Male vs. Female (simulated) mortalities")
}
\keyword{data}
