% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdealPeakFit.R
\name{selectNSeasons}
\alias{selectNSeasons}
\title{Identify the ideal number of seasons for a set of mortality data}
\usage{
selectNSeasons(T, max.season = NULL)
}
\arguments{
\item{T}{set of Surv objects representing time of death or censorship}

\item{max.season}{maximum number of seasons to survey}
}
\value{
a cmfitlist object comparing the various fits
}
\description{
Obtain AIC values for parameter estimates for a number of different models, based on the number of predicted seasons for a set of survival objects. This method will identify the most likely number of "mortality seasons" in one period for a species.
}
\examples{
T.morts1 <- simPeriodicMorts(1000, period = 1, 
                             meanhazard = 0.3, 
                             peaks = c(0.25, 0.75), 
                             durations = c(0.2, 0.1), 
                             weights = 0.3, 
                             plotme = FALSE)

fits = selectNSeasons(T.morts1)

print(fits)
}
