% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPeriodicMorts.R
\name{simPeriodicMorts}
\alias{simPeriodicMorts}
\title{Simulate periodic mortality process}
\usage{
simPeriodicMorts(n, A = 0.01, peaks = c(0.25, 0.75), rhos = c(0.6,
  0.6), weights = c(0.5), fixedCensoring = FALSE, period = 1,
  dt = 0.01, max.periods = 10, plotme = TRUE)
}
\description{
Simulate periodic mortality process
}
\examples{
par(oma = c(2,0,2,0));
T.morts1 <- simPeriodicMorts(300, period = 1, 
                             peaks = c(0.25, 0.75), 
                             rhos = c(0.8, 0.5), 
                             weights = 0.3, dt = .01, 
                             A = .02, plotme = TRUE);

with(attributes(T.morts1),
     title(paste0("A: ", A, "; peaks: ",
                  paste(peaks, collapse = ",")), outer = TRUE));

hist(T.morts1[,1] - floor(T.morts1)[,1], breaks = 20)
}
