% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmfactorfit.R
\name{summary.cmfactorfit}
\alias{summary.cmfactorfit}
\title{Produce a table explaining the results of a cmfactorfit model}
\usage{
\method{summary}{cmfactorfit}(x)
}
\arguments{
\item{x}{a cmfactorfit object}
}
\value{
a table comparing log-likelihood and AIC between null and multi-factor model, as well as a p-value from likelihood ratio test
}
\description{
Produce a table explaining the results of a cmfactorfit model
}
