% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmfit.R
\name{summary.cmfitlist}
\alias{summary.cmfitlist}
\title{Provides a short summary of the cmfitlist object, displaying the effectiveness of each model}
\usage{
\method{summary}{cmfitlist}(x)
}
\arguments{
\item{x}{a cmfitlist object}
}
\value{
a data frame describing the AIC, log-likelihood, number of parameters and parameter estimates for each model
}
\description{
Provides a short summary of the cmfitlist object, displaying the effectiveness of each model
}
\examples{
T.morts1 <- simPeriodicMorts(1000, period = 1, 
                             meanhazard = 0.3, 
                             peaks = c(0.25, 0.75), 
                             durations = c(0.2, 0.1), 
                             weights = 0.3, 
                             plotme = FALSE)

fits = selectNSeasons(T.morts1)

print(fits)
}
