% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.r
\docType{data}
\name{wah_morts}
\alias{wah_morts}
\title{Mortality data for Western Arctic Herd Caribou}
\format{An object of class \code{data.frame} with 90 rows and 4 columns.}
\source{
U.S. National Park Service, Alaska
}
\usage{
data(wah_morts)
}
\description{
Describing the data.
}
\examples{
data(wah_morts)
require(ggplot2)
ggplot(wah_morts \%>\% arrange(start),
aes(x = start, y = id, col = Fate)) + 
  geom_errorbarh(aes(xmin = start, xmax = end))

}
\keyword{data}
