% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappedCauchyFunctions.R
\name{wc}
\alias{wc}
\alias{iwc}
\alias{mwc}
\alias{imwc}
\title{Wrapped Cauchy and Integrated Wrapped Cauchy functions}
\usage{
wc(t, mu, rho, tau)

iwc(t, mu, rho, tau)

mwc(t, mus, rhos, omegas, tau)

imwc(t, mus, rhos, omegas, tau)
}
\arguments{
\item{t}{time}

\item{mu}{mean peak}

\item{rho}{concentration parameter}

\item{tau}{period}
}
\description{
Fundamental periodic hazard function and its (analytical) integral - required for efficient estimation. These are mainly internal functions. They are both parameterized in terms of peak mean  \eqn{\mu}, concentration parameter  \eqn{\rho}, and period \eqn{\tau}. They are "unweighted", i.e. \deqn{\int_0^\tau f(t) dt = \tau}
}
\examples{

# distribution function
curve(dwc(x, mu = 100, rho = .7, tau = 365), xlim = c(0,365), n = 1e4, ylab = "hazard", xlab = "time")
curve(dwc(x, mu = 100, rho = .5, tau = 365), add = TRUE, col = 2)
curve(dwc(x, mu = 100, rho = .3, tau = 365), add = TRUE, col = 3)


# compare numerical and analytical integral

require(microbenchmark)
microbenchmark(
  I1 <- integrate(dwc, mu =  100, rho = .1, tau = 365,  lower = 0, upper = 1000)$value, 
  I2 <- iwc(t = 1000, mu =  100, rho = .1, tau = 365)
)

I1
I2
}
