% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_cmfit.R
\name{AIC.cmfit}
\alias{AIC.cmfit}
\title{Determine AIC for a cmfit object}
\usage{
\method{AIC}{cmfit}(x)
}
\arguments{
\item{x}{a cmfit object}
}
\value{
AIC (a measure of model accuracy) for the cmfit object
}
\description{
Determine AIC for a cmfit object
}
\examples{
T.morts1 <- simPeriodicMorts(1000, period = 365, 
                             meanhazard = 0.3 / 365, 
                             peaks = c(0.25 * 365, 0.75 * 365), 
                             durations = c(0.3 * 365, 0.1 * 365), 
                             weights = 0.7, 
                             plotme = FALSE)

##MLE for parameters based on simulated data
fits = fit_cyclomort(T.morts1, n.seasons = 2)
print(fits)
plot(fits)
##Actual parameter values from simulated data
attributes(T.morts1)
##Measures of model accuracy
AIC(fits)
logLik(fits)
}
