% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCycloSurv.R
\name{createCycloSurv}
\alias{createCycloSurv}
\title{Create an object that can be fitted with a given period from survival (time of death or censoring) data}
\usage{
createCycloSurv(start, end, event, t0 = NULL, period,
  timeunits = "days")
}
\arguments{
\item{start}{a vector measuring time of birth (as a Date)}

\item{end}{a vector measuring time of death or censoring (as a Date)}

\item{event}{a vector of booleans (0 = alive, 1 = dead) detailing the status of the observation}

\item{t0}{start time for all times (start and end).  Note, the selection of t0 is related to the phase shift of the periodic analysis and should be set to the period of lowest hazard as determined, e.g., from a visual analysis.}

\item{period}{length of one period in the input data}
}
\value{
a Surv object with an attribute "period" that reads the period.
}
\description{
Create an object that can be fitted with a given period from survival (time of death or censoring) data
}
