% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCycloSurv.R
\name{createCycloSurv}
\alias{createCycloSurv}
\title{Create an object that can be fitted with a given period from survival (time of death or censoring) data}
\usage{
createCycloSurv(start, end, event, t0 = NULL, period,
  timeunits = "days")
}
\arguments{
\item{start}{a vector measuring time of birth (as a Date)}

\item{end}{a vector measuring time of death or censoring (as a Date)}

\item{event}{a vector of booleans (0 = alive, 1 = dead) detailing the status of the observation}

\item{t0}{reference time for event times.  By default, \code{t0} is set to January 1 of the first year of observations (if times are POSIX).}

\item{period}{length of one period in the input data}
}
\value{
a Surv object with an attribute "period" that reads the period.
}
\description{
Create an object that can be fitted with a given period from survival (time of death or censoring) data
}
