\name{cyclomort-package}
\alias{cyclomort-package}
\alias{cyclomort}
\docType{package}
\title{
\packageTitle{cyclomort}
}
\description{
\packageDescription{cyclomort}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cyclomort}
\packageIndices{cyclomort}
With the functions included in this package one can fit mortality data (stored as Surv objects) to a model that assumes a hazard function with a mixed wrapped Cauchy distribution. We use maximum likelihood estimation to produce esimates for the peaks in mortality over the course of the year as well as the strength or condensation of mortality around these peak indices. We also include a function that simulates a random set of mortalities given a set of parameters, allowing one to test the accuracy of their fits.
}
\author{
\packageAuthor{cyclomort}

Maintainer: \packageMaintainer{cyclomort}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
