% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclomort.R
\docType{package}
\name{cyclomort-package}
\alias{cyclomort}
\alias{cyclomort-package}
\title{Cyclomort: a package for parameteric periodic survival modeling}
\description{
With the functions included in this package one can fit mortality data (stored as Surv objects) to a model that assumes a hazard function with a mixed wrapped Cauchy distribution. We use maximum likelihood estimation to produce esimates for the peaks in mortality over the course of the year as well as the strength or condensation of mortality around these peak indices. We also include a function that simulates a random set of mortalities given a set of parameters, allowing one to test the accuracy of their fits.
}
\author{
\strong{Maintainer}: Eliezer Gurarie \email{egurarie@umd.edu}

Other contributors:
\itemize{
  \item Thompson Peter [contributor]
}

}
