% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRhoDelta.R
\name{findDelta}
\alias{findDelta}
\alias{findRho}
\title{Convert Rho to Tau}
\usage{
findDelta(rho)

findRho(delta)
}
\description{
Functions for converting the concentration parameter rho to the season duration parameter delta and vice versa.  They are: \code{findDelta(rho)} and \code{findDelta(delta)}.  These are not very exciting functions, but making them work correctly involved quite a bit of calculus and algebra!
}
\examples{
findDelta(rho = 0.9); findRho(0.0167)
findDelta(rho = 0.1); findRho(0.218)


# Plot the relationship
par(mfrow = c(1,2))
rhos <- seq(0, 1, length = 1e3)
plot(rhos, Vectorize(findDelta, vectorize.args = "rho")(rhos), ylab = "deltas", type = "l")
deltas <- seq(0, .5, length = 1e3)
plot(deltas, Vectorize(findRho, vectorize.args = "delta")(deltas), ylab = "rhos", type = "l")
}
