% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cyclomort.R
\name{fit_cyclomort}
\alias{fit_cyclomort}
\title{Get maximum likelihood estimates for periodic mortality data. Uses the optim
function along with a named vector of initial guesses and returns the parameter
estimates.}
\usage{
fit_cyclomort(T, inits = NULL, n.seasons = 2)
}
\arguments{
\item{T}{set of Surv objects representing time of death or censorship}

\item{n.seasons}{expected number of seasons if p0 is not entirely filled out}

\item{p0}{set of initial guesses; a named vector or list with values for "peak" and "duration". Leaving some or all of these parameters as NULL will trigger the automatic selection of an initial guess.}
}
\value{
parameter estimates for weights, rhos, peaks and A
}
\description{
Get maximum likelihood estimates for periodic mortality data. Uses the optim
function along with a named vector of initial guesses and returns the parameter
estimates.
}
\examples{
T.morts1 <- simPeriodicMorts(1000, period = 1, 
                             meanhazard = 0.3, 
                             peaks = c(0.25, 0.75), 
                             durations = c(0.3, 0.1), 
                             weights = 0.7, 
                             plotme = FALSE)

p0 = c(meanhazard = 0.05, 
       peak1 = 0.25, peak2 = 0.75, 
       duration1 = 0.2, duration2 = 0.2, 
       weight1 = 0.5, weight2 = 0.5)

##MLE for parameters based on simulated data
fits = fit_cyclomort(T.morts1, n.seasons = 2)
print(fits)
plot(fits)
##Actual parameter values from simulated data
attributes(T.morts1)
##Measures of model accuracy
AIC(fits)
logLik(fits)



T <- T.morts1 
T <- T[T[,1] > 0]
}
