% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cyclomort.R
\name{fit_cyclomort}
\alias{fit_cyclomort}
\title{Get maximum likelihood estimates for periodic mortality data. Uses the optim
function along with a named vector of initial guesses and returns the parameter
estimates.}
\usage{
fit_cyclomort(T, inits = NULL, n.seasons = 2, method = "L-BFGS-B",
  period = NULL)
}
\arguments{
\item{T}{a cycloSurv object recording start and end times as well as status (dead/censored) and the length of one full period}

\item{inits}{set of initial guesses; a named vector or list with values for "peak" and "duration". Leaving some or all of these parameters as NULL will trigger the automatic selection of an initial guess.}

\item{n.seasons}{number of seasons to fit model to}

\item{method}{method for optim call}

\item{period}{expected periodicity of survival data. Can be passed in with cycloSurv input parameter.}
}
\value{
parameter estimates for weights, rhos, peaks and A
}
\description{
Get maximum likelihood estimates for periodic mortality data. Uses the optim
function along with a named vector of initial guesses and returns the parameter
estimates.
}
\examples{
T.morts1 <- simulate_cycloMort(1000, period = 365, 
                             meanhazard = 0.3 / 365, 
                             peaks = c(0.25 * 365, 0.75 * 365), 
                             durations = c(0.3 * 365, 0.1 * 365), 
                             weights = 0.7, 
                             plotme = FALSE)

##MLE for parameters based on simulated data
fits = fit_cyclomort(T.morts1, n.seasons = 2)
print(fits)
plot(fits)
##Actual parameter values from simulated data
attributes(T.morts1)
##Measures of model accuracy
AIC(fits)
BIC(fits)
logLik(fits)
}
