% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCIs.R
\name{getCIs}
\alias{getCIs}
\title{Get 95% confidence intervals for each parameter that is estimated by the MLE}
\usage{
getCIs(fit)
}
\arguments{
\item{fit}{returned value from call to "optim" containing parameter estimates and Hessian matrix}
}
\value{
confidence intervals for each parameter in the estimate
}
\description{
Get 95% confidence intervals for each parameter that is estimated by the MLE
}
\examples{
T.morts1 <- simPeriodicMorts(300, period = 1, 
                             peaks = c(0.25, 0.75), 
                             rhos = c(0.8, 0.5), 
                             weights = 0.3, dt = .01, 
                             A = .02, plotme = FALSE)

p0 = c(A = 0.05, peak1 = 0.25, peak2 = 0.75, rho1 = 0.5, rho2 = 0.5, weight1 = 0.5)
##MLE for parameters based on simulated data
fits = fit_cyclomort(T.morts1, p0, dt = .01, period = 1)
getCIs(fits)
##Actual parameter values from simulated data
attributes(T.morts1)
}
