% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{loglike_optim}
\alias{loglike_optim}
\title{Log-likelihood function that is useable for the optim command}
\usage{
loglike_optim(pars, T, dt, period)
}
\arguments{
\item{pars}{named vector including "A", "peak", "weight" and "rho" names}

\item{T}{times of death or censoring as Surv objects}

\item{dt}{interval for plots as well as precision of random samples}

\item{period}{period of one mortality cycle}
}
\value{
likelihood value given named vector of parameters as well as set of observations
}
\description{
Log-likelihood function that is useable for the optim command
}
