% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.r
\docType{data}
\name{nwt_morts}
\alias{nwt_morts}
\title{Mortality data for Northwest territory boreal woodland caribou.}
\format{An object of class \code{data.frame} with 370 rows and 5 columns.}
\source{
Government of Northwest Territories, Canada
}
\usage{
data(nwt_morts)
}
\description{
Mortality data for Northwest territory boreal woodland caribou.
}
\examples{
data(nwt_morts)
require(ggplot2)
ggplot(nwt_morts \%>\% arrange(start) \%>\% mutate(id = factor(id, levels = id)),
aes(x = start, y = id, col = status)) + 
  geom_errorbarh(aes(xmin = start, xmax = end))

}
\keyword{data}
