% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmfit.R
\name{plot.cmfit}
\alias{plot.cmfit}
\title{Plot cmfit objects}
\usage{
\method{plot}{cmfit}(x, CI = TRUE, nreps = 10000, ...)
}
\arguments{
\item{x}{a cmfit object}
}
\value{
a plot comparing the estimated mortality curve (based on parameter estimates)
and the actual results (as a histogram).
}
\description{
Plot cmfit objects
}
\examples{
T.morts1 <- simPeriodicMorts(1000, period = 365, 
                             meanhazard = 0.3 / 365, 
                             peaks = c(0.25 * 365, 0.75 * 365), 
                             durations = c(0.3 * 365, 0.1 * 365), 
                             weights = 0.7, 
                             plotme = FALSE)

##MLE for parameters based on simulated data
fits = fit_cyclomort(T.morts1, n.seasons = 2)
print(fits)
plot(fits, type = "mortality")
##Actual parameter values from simulated data
attributes(T.morts1)
##Measures of model accuracy
AIC(fits)
logLik(fits)
}
