% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cmfit.R
\name{predict.cmfit}
\alias{predict.cmfit}
\title{Prediction method for cyclomort fits}
\usage{
\method{predict}{cmfit}(x, t = seq(0, x$period, length = 100),
  type = "hazard", CI = FALSE, CI.level = 0.05, nreps = 10000)
}
\arguments{
\item{x}{a cmfit object}

\item{t}{times for prediction}

\item{type}{either "hazard" or "timetoevent" - dictates what function is called}

\item{CI}{whether or not to compute 95\% confidence intervals}

\item{CI.level}{confidence level (must be on (0, 0.5) interval) for CIs if CI is TRUE}

\item{nreps}{number of samples drawn to generate confidence intervals.  The default 10^4 is more than enough, but is a little sluggish. 10^3 gives slightly rougher intervals, but is very fast}
}
\description{
Obtain predictions and - optionally - confidence intervals for the hazard function from a fitted periodic hazard functon. Confidence intervals are produced by assuming that the parameters follow a multivariate normal distribution
}
\examples{
T <- simPeriodicMorts(300, period = 1, peaks = c(0.3, 0.8), 
                      durations = c(0.15, 0.20), weights = c(3, 2)/5, 
                      meanhazard = 1, plotme = FALSE, max.periods = 6)
T.fit <- fit_cyclomort(T, n.seasons = 2)
predictions = predict(T.fit, t = 0.5, CI = TRUE, type = "hazard")

plot(T.fit)
}
