% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cmfitlist.R
\name{print.cmfitlist}
\alias{print.cmfitlist}
\title{Display cmfitlist objects as text}
\usage{
\method{print}{cmfitlist}(x)
}
\arguments{
\item{x}{a cmfitlist object}
}
\value{
a table comparing the different periodic mortality models, comparing them using AIC as well as the parameter estimates
}
\description{
Display cmfitlist objects as text
}
\examples{
T.morts1 <- simPeriodicMorts(1000, period = 1, 
                             meanhazard = 0.3, 
                             peaks = c(0.25, 0.75), 
                             durations = c(0.2, 0.1), 
                             weights = 0.3, 
                             plotme = FALSE)

fits = selectNSeasons(T.morts1, max.season = 4)

print(fits)
}
