% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.r
\docType{data}
\name{seasonalsex}
\alias{seasonalsex}
\title{Simulated data of seasonal mortality data for two sex groups}
\format{Simulated data of single-season mortalities for two sex groups:
\describe{
  \item{sex}{female (F) or male (M)}
  \item{event}{\code{cycloSurv} object of (censored) survival data}
}}
\usage{
data(seasonalsex)
}
\description{
See examples below for the process of simulating and visualizing these data using \code{\link{simulate_cycloSurv}}, and an example of analyzing these data with \code{\link{factorfit_cyclomort}}.
}
\examples{
# useful packages
require(ggplot2); require(magrittr); require(plyr)

# Example of simulating multi-factor data:
## Not run:** 
  n <- 100
  T.male = simulate_cycloSurv(n, period = 1, meanhazard = 0.3, peaks = .25, durations = .3)
  T.female = simulate_cycloSurv(n, period = 1, meanhazard = 0.3, peaks = .75, durations = .3)
  T <- with(rbind(T.male, T.female) \%>\% data.frame,
            create_cycloSurv(start = start, end = stop, 
                            event = status, period = 1))
  seasonalsex <- data.frame( sex = rep(c("M","F"), each = n), T = T)
## End(**Not run**)

# load and visualize simulated sex-specific survival data
data("seasonalsex")

seasonsex.df <- cbind(seasonalsex, as.matrix(seasonalsex$event) \%>\% as.data.frame) \%>\%
  arrange(sex,stop) \%>\% mutate(id = 1:length(start) \%>\% factor, 
                               status = c("Dead", "Censored")[2-status])

require(ggplot2)
ggplot(seasonsex.df, aes(x = start, y = id, col = sex)) + 
  geom_errorbarh(aes(xmin = start, xmax = stop, lty = status)) + 
    ggtitle("Simualted sex-specific mortality data")

ss.dead <- seasonsex.df \%>\% subset(status == "Dead") \%>\% mutate(time = (stop - floor(stop))) 
ggplot(ss.dead, aes(time, y=..density.., fill = sex)) + geom_density(alpha = 0.1) + 
  geom_histogram(alpha=0.5, position="identity")  + 
    ggtitle("Male vs. Female (simulated) mortalities")


# test differences

sex.fit <- factorfit_cyclomort(event ~ sex, data = seasonalsex, n.seasons = 1)
summary(sex.fit)
plot(sex.fit, ymax = 1.3)
}
\keyword{data}
