% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdealPeakFit.R
\name{selectNSeasons}
\alias{selectNSeasons}
\title{Identify the ideal number of seasons for a set of mortality data}
\usage{
selectNSeasons(T, dt = 0.01)
}
\arguments{
\item{T}{set of Surv objects representing time of death or censorship}

\item{dt}{interval for plots as well as precision of random samples}
}
\value{
a cmfitlist object comparing the various fits
}
\description{
Obtain AIC values for the MLEs for a five different mortality models, each one having a different number of seasons.
Decide which model should be used based on the lowest AIC; this will estimate how many
seasons in mortality occur during a periodic cycle for the given species.
}
\examples{
T.morts1 <- simPeriodicMorts(300, period = 1, 
                             peaks = c(0.25, 0.75), 
                             rhos = c(0.7, 0.7), 
                             weights = 0.2, dt = .01, 
                             A = .02, plotme = FALSE)

fits = getIdealPeakFit(T = T.morts1, dt = 0.01)
print(fits)
}
