% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdealPeakFit.R
\name{selectNSeasons}
\alias{selectNSeasons}
\title{Identify the ideal number of seasons for a set of mortality data}
\usage{
selectNSeasons(T, max.season = NULL)
}
\arguments{
\item{T}{set of Surv objects representing time of death or censorship}
}
\value{
a cmfitlist object comparing the various fits
}
\description{
Obtain AIC values for the MLEs for a five different mortality models, each one having a different number of seasons.
Decide which model should be used based on the lowest AIC; this will estimate how many
seasons in mortality occur during a periodic cycle for the given species.
}
\examples{
T.morts1 <- simPeriodicMorts(1000, period = 1, 
                             meanhazard = 0.3, 
                             peaks = c(0.25, 0.75), 
                             durations = c(0.2, 0.1), 
                             weights = 0.3, 
                             plotme = FALSE)

fits = selectNSeasons(T.morts1)

print(fits)
}
