% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmfactorfit.R
\name{summary.cmfactorfit}
\alias{summary.cmfactorfit}
\title{Produce a table explaining the results of a cmfactorfit model}
\usage{
\method{summary}{cmfactorfit}(x, coefs = FALSE)
}
\arguments{
\item{x}{a cmfactorfit object}

\item{coefs}{a boolean variable determining whether the individual summaries of each model component will be displayed along with the LRT resul}
}
\value{
a table comparing log-likelihood and AIC between null and multi-factor model, as well as a p-value from likelihood ratio test, or a list of such a table combined with the individual model summaries if coefs is listed as TRUE
}
\description{
Produce a table explaining the results of a cmfactorfit model
}
\examples{
# fit factorial model
data(seasonalsex)
x <- factorfit_cyclomort(event ~ sex, data = seasonalsex, n.seasons = 1)

# summary
summary(x, coefs = TRUE)
plot(x)

}
