% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmfit.R
\name{summary.cmfitlist}
\alias{summary.cmfitlist}
\title{Provides a short summary of the cmfitlist object, displaying the effectiveness of each model}
\usage{
\method{summary}{cmfitlist}(x)
}
\arguments{
\item{x}{a cmfitlist object}
}
\value{
a data frame describing the AIC, log-likelihood, number of parameters and parameter estimates for each model
}
\description{
Provides a short summary of the cmfitlist object, displaying the effectiveness of each model
}
\examples{
T.morts1 <- simPeriodicMorts(300, period = 1, 
                             peaks = c(0.25, 0.75), 
                             rhos = c(0.7, 0.7), 
                             weights = 0.2, dt = .01, 
                             A = .02, plotme = FALSE)

fits = getIdealPeakFit(T = T.morts1, dt = 0.01)
print(fits)
}
