% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.r
\docType{data}
\name{wah_morts}
\alias{wah_morts}
\title{Mortality data for Western Arctic Herd Caribou}
\format{Data frame with 171 rows and the following columns:
\describe{
  \item{id}{ID of animal}
  \item{start}{Date of beginning of collaring}
  \item{end}{Date of death or censoring}
  \item{Fate}{One of "DEAD", of (censored): "DROPOFF", "CollarFail", "LIVE ANIMAL"}
}}
\source{
U.S. National Park Service, Alaska
}
\usage{
data(wah_morts)
}
\description{
Describing the data.
}
\examples{
data(wah_morts)
require(ggplot2)
ggplot(wah_morts \%>\% arrange(start),
aes(x = start, y = id, col = Fate)) + 
  geom_errorbarh(aes(xmin = start, xmax = end))

}
\keyword{data}
